Debug the reset command,fix enemy sprite display, fix the problem in enemy moving behaviour, update the sprite sheet used in UseItem,

+++ Game1/Sprite_/Moving Sprite/UpMovingLinkSprite.cs
@@ -28,7 +28,7 @@ namespace Game1
             GlobalDefinitions.Position.Y = GlobalDefinitions.Position.Y - (float)5; //change the y axis position of Link
             if (GlobalDefinitions.Position.Y == 0)
             {
-                GlobalDefinitions.Position.Y = (float)GlobalDefinitions.GraphicsHeight;
+                GlobalDefinitions.Position.Y = GlobalDefinitions.GraphicsHeight;
             }
 
         }
